requirements.txt
python-nmap
requests
dnspython
